requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.22
watchfiles

# Database
sqlmodel==0.0.31
psycopg2-binary==2.9.11
alembic==1.18.3

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic[email]

# Environment
python-dotenv==1.2.1

# Time Series (for TimescaleDB operations)
pandas==3.0.0
numpy==2.4.1
